Scraper is almost complete #7 #1
The general scraper class is almost complete, tested on WITS offer
demand as well as EMI data. iPython seems to spit the dummy a little
bit when dealing with the compressions and I still need to write
a lot of seed instances in the config but this is almost complete
enough to replace three files. EMIScraper, wits_scraper and nodal_demand
scrapers. Still need to write a lot of docstrings.

One caveat is that now a few edge cases are handled by string checking.
I'm not too concerned about this, but it does mean some of the code
has different control loops depending upon which string is parsed
which can make it more difficult to reason about.

Overall, I think this is a good addition.
Will need to lookm over the POCP/Wits5min scrapers as well
to see how they'll fit into all of this