Scraper Class Completed, Fixes #1 #7
The Scraper Class has been completed and should work across both
the WITS and the EMI website. Can be extended for ASX and the Fivemin Scraper
as well when I have a bit of time to look into this further.

Key points, the class works by providing a seed of information
which changes the pathway the scraper will hit.
As we're dealing with other peoples websites and cannot control them
this can get really messy.