Interface to DB queries have been written.
This is a major upgrade which Fixes #16 and Fixes #22.
can now use the "query_offer" method to automatically query from
the DB tables and return results as OfferFrames.

No need to worry about where the Data is coming from.
It should "just" work. I've specified that a date range is required or it
will try to query everything which is about 2GB of data per master table.
Not really sustainable.

I'm quite excited by this. I still need to write some documentation and
there is some very basic type checking.