Working on documentation and getting all of the SQL queries for the Database sorted out.
Can hopefully change this from using the default class to a more functional style by
the end of the day which should make testing and maintaining it a lot easier. Still need
to work out the data import functionality but the queries are getting significantly
easier to work with at the very least.