Queries are now working for both prices and demand outside of the core NZEMDB class which is nice
Still need to do the offers but I'm going to leave this for when I have some more energy.
Can then move on to the data import functionality and begin to add some new features once this
fucking rewrite is over. I should also write a lot more tests but that can wait for now.

I've also fixed some bugs which were arising so it should be more stable overall to query the results